requirements.txt
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
filelock==3.17.0
flasgger==0.9.7.1
Flask==3.1.0
fsspec==2025.2.0
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.1.1
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==24.2
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.22.3
safetensors==0.5.2
six==1.17.0
sympy==1.13.3
tokenizers==0.21.0
torch==2.2.2
tqdm==4.67.1
transformers==4.48.2
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0
